Большой фикс
-Переделан алгоритм подсчета энергий. Теперь энергия состояния
автоматически вычисляется с учетом разности от основного состояния, а
минимум и максимум не вычисляются, если ранее уже были посчитаны.
-Добавлен автоматический метод вычисления PartArray::E()
-Исправлен баг с ИДом частицы
-Удалены методы calcEnergy1 и calcEnergy2, которые ничем не отличались.
Вместо них теперь реализован метод EComplete, который не меняет
состояния системы при своей работе.
-Добавлен PartArray::ECompleteFast(), который оптимизирован для
вычисления энергий больших систем
-PartArray::getById() теперь возвращает 0 при отсутствии найденной
частицы
-PartArray::dropRandom(), PartArray::dropChain() и
PartArray::dropSpinIce() перед генерацией очищает массив частиц
-Добавлены события ChangeState и ChangeSystem, которые вызываются при
изменении состояния и изменении системы соответственно.
-При сохранении и загрузке системы из файла, добавлено сохранение
минимального и максимального состояний
-В класс StateMachineFree добавлен метод очистки clear()
-Обновлены тестовые наборы

+++ Part.h
@@ -41,7 +41,7 @@ public:
     int Id();
 
 protected:
-    unsigned int id;
+    long int id;
     Part(unsigned int id); //конструктор с установленным ИДом разрешен только для дружественных классов
 };
